Convert MMC to using CAM. This will allow sdio drivers to be written more easily.
The MMCNULL driver is the emulated MMC controller. Having such module
is supposed to help in developing CAM version of the MMC stack.

--- remove before commit - notes to Ilya and reviewer ---

This commit is all the kernel parts of Ilya's mmccam branch collapsed
into one big commit. I've added the mmccam option later in this series
and will make sure we can build w/o it, but that's for another day.

I also plan to separate out the mmcnull driver out for the commit.  It
would be more useful if we could map a file into the driver somehow so
we could emulate things better for testing w/o actual MMC hardware.

There's a question about sdhci_start_slot vs sdhci_cam_start_slot.
And this might not compile since upstream changed cmd to a pointer.