Add sdio(4) peripheral device
This device will attach to SDIO cards and provide a bus-like interface to
its children. Also it will read CCCR / FBR registers of the card
and get vendor and product IDs so that child devices are able to
properly do probing.

sdio.ko is dynamically unloadable, but nobody has been unloading peripheral drivers before. So the code might not be quite stable.

Also add brcmwl driver as example of SDIO device driver.